pyproject.toml
[project]
name = "documentassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.1.3",
    "chromadb>=1.3.5",
    "docx>=0.2.4",
    "fastapi[all]>=0.122.0",
    "langchain[all,experimental]>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "passlib[bcrypt]==1.7.4",
    "pdfplumber>=0.11.8",
    "phonenumbers==9.0.17",
    "psycopg2-binary>=2.9.11",
    "pydantic-extra-types>=2.10.6",
    "pyjwt>=2.10.1",
    "pymupdf>=1.26.6",
    "python-docx>=1.2.0",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.19",
    "sentence-transformers>=5.1.2",
    "sqlalchemy[asyncio]==2.0.36",
    "uvicorn[standard]==0.30.1",
    "vertexai>=1.43.0",
    "langchain-google-vertexai>=3.1.0",
    "generativeai>=0.0.1",
    "google-generativeai>=0.8.5",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.0",
    "langchain-huggingface>=1.1.0",
    "stripe>=14.0.1",
]
